+++ cognite/powerops/resync/config/market/dayahead.py
@@ -42,9 +42,9 @@ class BidProcessConfig(Configuration):
     @field_validator("price_scenarios", mode="after")
     def ensure_no_duplicates(cls, value: list[PriceScenarioID], info: FieldValidationInfo):
         scenario_ids = Counter(scenario.id for scenario in value)
-        if duplicates := [
-            scenario_id for scenario_id, count in scenario_ids.items() if count > 1
-        ]:
+        if duplicates := [scenario_id for scenario_id, count in scenario_ids.items() if count > 1]:
             bidprocess_name = info.data.get("name", "unknown")
-            raise ValueError(f"Duplicate price scenarios for bidprocess {bidprocess_name} was found: {list(duplicates)}")
+            raise ValueError(
+                f"Duplicate price scenarios for bidprocess {bidprocess_name} was found: {list(duplicates)}"
+            )
         return value