+++ tests/test_integration/resync/test_plan.py
@@ -37,6 +37,7 @@ def plan_test_cases():
             id=f"{test_case['data_path']} {test_case['model_name']}",
         )
 
+
 @pytest.mark.skip("Temporaryly skip")
 @pytest.mark.parametrize("data_path, market, model_name, dump_folder, config_file", list(plan_test_cases()))
 def test_plan(